+++ server/routes/cycle.js
@@ -1,17 +1,16 @@
 const { Router } = require('express');
-const CycleController = require('../data/controllers/cycleController');
+const CycleController = require('../controllers/CycleController');
 
 const router = Router();
 
+// Add or delete a cycle
 router.post('/', CycleController.configureCycle);
+router.delete('/', CycleController.deleteCycle);
 
-router.post('/delete', CycleController.deleteCycle);
+// Gets all cycles
+router.get('/', CycleController.getAllCycles);
 
-// Maybe I should use a "get" like: router.get('/', CycleController.getAllCycles);
-// and then when you want to get all cycles, you send a GET request, when you want to save, you send a POST?
-router.post('/cycles', CycleController.getAllCycles);
-
-// This might not belong here
-router.post('/testversions', CycleController.getAllTestVersions);
+// Gets test version information necessary to initiate cycle
+router.get('/testversions', CycleController.getAllTestVersions);
 
 module.exports = router;
